chore: adding AMF cache option to optimize build time

+++ ci-app/apic/test-runners/amf-compiler.sh
@@ -14,9 +14,6 @@ AMF_DIR="${BUILD_DIR}/amf/"
 echo "Building AMF library in $AMF_DIR"
 
 mkdir ${AMF_DIR}
-mkdir "${AMF_DIR}/src/"
-mkdir "${AMF_DIR}/lib/"
-mkdir "${AMF_DIR}/lib/bin"
 cd $AMF_DIR
 
 git clone --depth=50 --branch=$BRANCH https://github.com/aml-org/amf.git src
@@ -28,18 +25,36 @@ else
   git checkout ${COMMIT_SHA}
 fi
 
-echo "Building AMF client library."
-sbt clientJS/fullOptJS
-echo "Executing buildjs.sh from AMF project"
-./amf-client/js/build-scripts/buildjs.sh
+SHA=`git rev-parse HEAD`
+CACHED_NAME="$SHA.tar.gz"
+CACHED="$HOME/amf-cache/${CACHED_NAME}"
 
-echo "Copying AMF library to the working location."
-cd ../
+if test -f "$CACHED"; then
+  cd ../
+  tar -xzf $CACHED
+else
+  echo "Building AMF client library."
+  sbt clientJS/fullOptJS
+  echo "Executing buildjs.sh from AMF project"
+  ./amf-client/js/build-scripts/buildjs.sh
 
-cp src/amf-client/js/amf.js lib/
-cp src/amf-client/js/package.json lib/
-cp src/amf-client/js/bin/amf lib/bin/
+  echo "Copying AMF library to the working location."
+  cd ../
 
-echo "Installing AMF project dependencies"
-cd lib/
-npm i
+  mkdir "${AMF_DIR}/lib/"
+  mkdir "${AMF_DIR}/lib/bin"
+
+  cp src/amf-client/js/amf.js lib/
+  cp src/amf-client/js/package.json lib/
+  cp src/amf-client/js/bin/amf lib/bin/
+
+  echo "Installing AMF project dependencies"
+  cd lib/
+  npm i
+  cd ../
+  echo "Caching commit to $CACHED"
+  tar -czf $CACHED_NAME lib
+  mkdir -p "$HOME/amf-cache/"
+  cp $CACHED_NAME $CACHED
+  rm $CACHED_NAME
+fi
